fix: hook 분리하여 빌드에러 해결
