+++ src/app/search/_components/SearchMobile.tsx
@@ -2,46 +2,10 @@
 
 import HeaderCategorys from "@/components/layout/HeaderCategorys";
 import SearchBar from "@/components/layout/SearchBar";
-import { useRouter, useSearchParams } from "next/navigation";
-import { useEffect, useRef, useState } from "react";
+import usePopupState from "@/lib/hooks/search/usePopupState";
 
 const SearchMobile = () => {
-  const searchParams = useSearchParams();
-  const router = useRouter();
-  const [isOpen, setIsOpen] = useState(false);
-
-  const bodyRef = useRef<HTMLElement | null>(null);
-
-  useEffect(() => {
-    const popup = searchParams.get("popup");
-    setIsOpen(popup === "true");
-  }, [searchParams]);
-
-  useEffect(() => {
-    if (!bodyRef.current) {
-      bodyRef.current = document.body;
-    }
-
-    if (isOpen) {
-      bodyRef.current.style.overflow = "hidden";
-      document.documentElement.style.overflow = "hidden";
-    } else {
-      bodyRef.current.style.overflow = "";
-      document.documentElement.style.overflow = "";
-    }
-
-    return () => {
-      if (bodyRef.current) {
-        bodyRef.current.style.overflow = "";
-        document.documentElement.style.overflow = "";
-      }
-    };
-  }, [isOpen]);
-
-  const handleClosePopup = () => {
-    setIsOpen(false);
-    router.replace("/search");
-  };
+  const { isOpen, closePopup } = usePopupState();
 
   return (
     <>
